+++ src/models/money.rs
@@ -35,28 +35,18 @@ use std::str::FromStr;
 /// let m1 = Money::from((eur.clone(), BigRational::from(BigInt::from(100))));
 /// let m2 = Money::from((eur.clone(), BigRational::from(BigInt::from(200))));
 /// # let m3 = Money::from((eur.clone(), BigRational::from(BigInt::from(300))));
-/// let b1 = m1.clone() + m2; // 300 euros
+/// let b1 = m1.clone() + m2.clone(); // 300 euros
 /// # assert_eq!(*b1.balance.get(&Some(eur.clone())).unwrap(), m3);
 ///
 /// // Multicurrency works as well
 /// let d1 = Money::from((usd.clone(), BigRational::from(BigInt::from(50))));
 /// let b2 = d1.clone() + m1.clone(); // 100 euros and 50 usd
-/// # assert_eq!(b2.balance.len(), 2);
-/// # assert_eq!(*b2.balance.get(&Some(eur.clone())).unwrap(), m1);
-/// # assert_eq!(*b2.balance.get(&Some(usd.clone())).unwrap(), d1);
+/// assert_eq!(b2.balance.len(), 2);
+/// assert_eq!(*b2.balance.get(&Some(eur.clone())).unwrap(), m1);
+/// assert_eq!(*b2.balance.get(&Some(usd.clone())).unwrap(), d1);
 ///
-/// // There are various display formats
-/// // -4_285_714.28571... EUR
-/// let mut euro = Currency::from("eur");
-/// euro.set_decimal_separator(',');
-/// euro.set_thousands_separator('.');
-/// euro.set_digit_grouping(DigitGrouping::Indian);
-/// let rc_euro = Rc::new(euro);
-/// let money = Money::from((rc_euro, BigRational::new(BigInt::from(-30000000), BigInt::from(7))));
-/// assert_eq!(format!("{}", &money), "-42.85.714,29 eur");
-/// assert_ne!(format!("{}", &money), "-4285714,29 eur");
-///
-///
+/// let b3 = b1 - Balance::from(m2.clone()) + Balance::from(Money::new());
+/// assert_eq!(b3.to_money().unwrap(), m1);
 /// ```
 #[derive(Clone, Debug, PartialOrd)]
 pub enum Money {
@@ -231,21 +221,20 @@ impl Display for Money {
             Money::Zero => write!(f, "{}", "0"),
             Money::Money { amount, currency } => {
                 // Suppose: -1.234.567,000358 EUR
+
+                // Get the format
+                let format = currency.display_format.borrow();
+
                 // num = trunc + fract
                 let base: i32 = 10;
                 let mut integer_part = amount.trunc(); // -1.234.567
 
-                // Read decimals from format, two as default
-                let decimals = match currency.get_precision() {
-                    Some(p) => p,
-                    None => {
-                        if amount.fract().is_zero() {
-                            0
-                        } else {
-                            2
-                        }
-                    }
+                // Read number of decimals from format
+                let decimals = match format.max_decimals {
+                    Some(d) => d,
+                    None => format.precision,
                 };
+
                 let decimal_part = (amount.fract() * BigInt::from(base.pow(decimals as u32 + 2)))
                     .abs()
                     .trunc();
@@ -271,19 +260,19 @@ impl Display for Money {
                             integer_part -= BigInt::from(1);
                         }
                     }
-                    let decimal_separator = currency.get_decimal_separator_str();
+                    let decimal_separator = format.get_decimal_separator_str();
                     decimal_str =
                         format!("{}{:0width$}", decimal_separator, decimal, width = decimals);
                 }
 
                 let integer_str = {
-                    match currency.get_digit_grouping() {
+                    match format.get_digit_grouping() {
                         DigitGrouping::None => integer_part.numer().abs().to_string(), // Do nothing
                         grouping => {
                             let mut group_size = 3;
                             let mut counter = 0;
                             let mut reversed = vec![];
-                            let thousands_separator = currency.get_thousands_separator_str();
+                            let thousands_separator = format.get_thousands_separator_str();
                             for c in integer_part.to_string().chars().rev() {
                                 if c == '-' {
                                     continue;
@@ -305,10 +294,10 @@ impl Display for Money {
                 };
 
                 let amount_str = format!("{}{}", integer_str, decimal_str);
-                match currency.symbol_placement {
+                match format.symbol_placement {
                     CurrencySymbolPlacement::BeforeAmount => {
                         if amount.is_negative() {
-                            match currency.negative_amount_display {
+                            match format.negative_amount_display {
                                 NegativeAmountDisplay::BeforeSymbolAndNumber => {
                                     write!(f, "-{}{}", currency.get_name(), amount_str)
                                 }
@@ -328,7 +317,7 @@ impl Display for Money {
                     }
                     CurrencySymbolPlacement::AfterAmount => {
                         if amount.is_negative() {
-                            match currency.negative_amount_display {
+                            match format.negative_amount_display {
                                 NegativeAmountDisplay::BeforeSymbolAndNumber
                                 | NegativeAmountDisplay::BeforeNumberBehindCurrency => {
                                     write!(f, "-{} {}", amount_str, currency.get_name())