+++ src/parser.rs
@@ -10,7 +10,8 @@ use std::collections::HashSet;
 use std::fs::read_to_string;
 use std::path::PathBuf;
 
-use crate::models::{Account, Comment, Currency, Payee, Transaction};
+use crate::models::{Account, Comment, Currency, HasName, Payee, Transaction};
+use crate::parser::utils::count_decimals;
 use crate::{models, CommonOpts, List};
 use pest::Parser;
 
@@ -151,6 +152,11 @@ impl<'a> Tokenizer<'a> {
                                 }
                                 Rule::commodity => {
                                     let commodity = self.parse_commodity(inner);
+                                    if let Ok(old_commodity) =
+                                        ledger.commodities.get(&commodity.get_name())
+                                    {
+                                        commodity.update_precision(old_commodity.get_precision());
+                                    }
                                     ledger.commodities.remove(&commodity);
                                     ledger.commodities.insert(commodity);
                                 }
@@ -173,19 +179,33 @@ impl<'a> Tokenizer<'a> {
                                 ];
                                 for (currency, format) in currencies {
                                     if let Some(c) = currency {
-                                        let found = ledger.commodities.get(c);
-                                        if found.is_err() {
-                                            if options.pedantic {
-                                                panic!("Error: commodity {} not declared.", c);
-                                            }
-                                            if options.strict {
-                                                eprintln!("Warning: commodity {} not declared.", c);
-                                            }
+                                        match ledger.commodities.get(c) {
+                                            Err(_) => {
+                                                if options.pedantic {
+                                                    panic!("Error: commodity {} not declared.", c);
+                                                }
+                                                if options.strict {
+                                                    eprintln!(
+                                                        "Warning: commodity {} not declared.",
+                                                        c
+                                                    );
+                                                }
 
-                                            let mut commodity = Currency::from(c.as_str());
-                                            commodity
-                                                .set_format(format.as_ref().unwrap().to_owned());
-                                            ledger.commodities.insert(commodity);
+                                                let commodity = Currency::from(c.as_str());
+                                                if let Some(format_string) = format {
+                                                    commodity.update_precision(count_decimals(
+                                                        format_string.as_str(),
+                                                    ));
+                                                }
+                                                ledger.commodities.insert(commodity);
+                                            }
+                                            Ok(c) => {
+                                                if let Some(format_string) = format {
+                                                    c.update_precision(count_decimals(
+                                                        format_string.as_str(),
+                                                    ));
+                                                }
+                                            }
                                         }
                                     }
                                 }
@@ -205,7 +225,7 @@ impl<'a> Tokenizer<'a> {
                 eprintln!("Error found in line {}", e)
             }
         }
-        // dbg!(&ledger);
+
         ledger
     }
 }